v1.3/requirements.txt
Pillow==9.5.0
Pillow==10.0.0
ttkthemes==3.2.2
